Adding support for *.jsx file extensions.
React now supports ES6 classes, as of React 0.13, however they usually use the `*.jsx` extension (as it adds it's own set of syntax which is compiled to regular Javascript). This tiny PR should (as far as I can tell) allow the language-javascript-better plugin for Atom to support `*.jsx` files with ES6 syntax checking.